Refactor cli as script
